--- libs/movie/src/lib/movie/components/display-list/display-list.component.html
@@ -1,46 +1,0 @@
-<table mat-table [dataSource]="dataSource" class="mat-elevation-z8">
-  <!-- Picture column -->
-  <ng-container matColumnDef="picture">
-    <th mat-header-cell *matHeaderCellDef></th>
-    <td mat-cell *matCellDef="let movie">
-      <img *ngIf="movie.poster as poster" [ref]="poster" asset="empty_poster.webp">
-    </td>
-  </ng-container>
-
-  <!-- Movie title column -->
-  <ng-container matColumnDef="title">
-    <th mat-header-cell *matHeaderCellDef>Name</th>
-    <td mat-cell *matCellDef="let movie">{{ movie.title.original }}</td>
-  </ng-container>
-
-  <!-- Director column -->
-  <ng-container matColumnDef="director">
-    <th mat-header-cell *matHeaderCellDef>Director</th>
-    <!-- We want only to show one director because of the design right now -->
-    <td mat-cell *matCellDef="let movie">
-      <span *ngIf="movie.directors.length">{{ movie.directors[0].firstName }} {{ movie.directors[0].lastName }}</span>
-    </td>
-  </ng-container>
-
-  <!-- Production year column -->
-  <ng-container matColumnDef="releaseYear">
-    <th mat-header-cell *matHeaderCellDef>Release Year</th>
-    <td mat-cell *matCellDef="let movie">{{ movie.releaseYear }}</td>
-  </ng-container>
-
-  <!-- Action column -->
-  <ng-container matColumnDef="action">
-    <th mat-header-cell *matHeaderCellDef></th>
-    <td mat-cell *matCellDef="let movie">
-      <ng-container *ngIf="(toggle$(movie.id) | async); else alreadyAdded">
-        <mat-icon (click)="removeFromWishlist(movie, $event)" svgIcon="favorite"></mat-icon>
-      </ng-container>
-      <ng-template #alreadyAdded>
-        <mat-icon (click)="addToWishlist(movie, $event)" svgIcon="favorite_border"></mat-icon>
-      </ng-template>
-    </td>
-  </ng-container>
-
-  <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-  <tr mat-row *matRowDef="let row; columns: displayedColumns" (click)="navigate.emit(row.id)"></tr>
-</table>
