--- libs/movie/src/lib/movie/components/display-list/display-list.component.scss
@@ -1,7 +1,0 @@
-table {
-  width: 100%;
-  img {
-    height: 40px;
-    padding-top: 10px;
-  }
-}
