--- libs/movie/src/lib/movie/components/display-list/display-list.component.ts
@@ -1,48 +1,0 @@
-import { Output, EventEmitter } from '@angular/core';
-import { ChangeDetectionStrategy } from '@angular/core';
-import { Movie } from './../../+state/movie.model';
-import { Component, Input } from '@angular/core';
-import { MatTableDataSource } from '@angular/material/table';
-import { CartService } from '@blockframes/cart/+state/cart.service';
-import { MatSnackBar } from '@angular/material/snack-bar';
-import { CartQuery } from '@blockframes/cart/+state/cart.query';
-
-@Component({
-  selector: '[movies] movie-display-list',
-  templateUrl: './display-list.component.html',
-  styleUrls: ['./display-list.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-export class MovieDisplayListComponent {
-  public displayedColumns: string[] = ['picture', 'title', 'director', 'releaseYear', 'action'];
-  public dataSource: MatTableDataSource<Movie>;
-
-  @Input()
-  set movies(movies: Movie[]) {
-    this.dataSource = new MatTableDataSource(movies);
-  }
-
-  @Output() navigate = new EventEmitter<string>();
-
-  constructor(
-    private cartService: CartService,
-    private snackbar: MatSnackBar,
-    private cartQuery: CartQuery,
-  ) {}
-
-  public toggle$(movieId: string) {
-    return this.cartQuery.isAddedToWishlist(movieId);
-  }
-
-  public addToWishlist(movie: Movie, event: Event) {
-    event.stopPropagation();
-    this.cartService.updateWishlist(movie);
-    this.snackbar.open(`${movie.title.international} has been added to your selection.`, 'close', { duration: 2000 });
-  }
-
-  public removeFromWishlist(movie: Movie, event: Event) {
-    event.stopPropagation();
-    this.cartService.updateWishlist(movie);
-    this.snackbar.open(`${movie.title.international} has been removed from your selection.`, 'close', { duration: 2000 });
-  }
-}
