+++ tsconfig.base.json
@@ -29,7 +29,6 @@
       "@blockframes/ethers/*": ["libs/ethers/src/lib/*"],
       "@blockframes/smart-contracts/*": ["smart-contracts/build/*"],
       "@blockframes/smart-contracts/quorum/*": ["smart-contracts/quorum/*"],
-      "@blockframes/cart/*": ["libs/organization/src/lib/cart/*"],
       "@blockframes/organization/*": ["libs/organization/src/lib/organization/*"],
       "@blockframes/organization/types": [
         "libs/organization/src/lib/organization/+state/organization.firestore.ts"
